+++ init.rb
@@ -5,38 +5,11 @@ Redmine::Plugin.register :additionals do
   name 'Additionals'
   author 'AlphaNodes GmbH'
   description 'Customizing Redmine, providing wiki macros and act as a library/function provider for other Redmine plugins'
-  version '2.0.1'
+  version '2.0.2'
   author_url 'https://alphanodes.com/'
   url 'https://github.com/alphanodes/additionals'
 
-  default_settings = {
-    account_login_bottom: '',
-    add_go_to_top: 0,
-    custom_help_url: 'http://www.redmine.org/guide',
-    disabled_modules: nil,
-    external_urls: 1,
-    global_footer: '',
-    global_sidebar: '',
-    wiki_pdf_header: '',
-    wiki_pdf_remove_title: 0,
-    wiki_pdf_remove_attachments: 0,
-    global_wiki_footer: '',
-    global_wiki_header: '',
-    global_wiki_sidebar: '',
-    issue_auto_assign_role: '',
-    issue_auto_assign_status: '',
-    issue_auto_assign: 0,
-    legacy_smiley_support: 0,
-    new_ticket_message: 'Don\'t forget to define acceptance criteria!',
-    overview_bottom: '',
-    overview_right: '',
-    overview_top: '',
-    project_overview_content: 'Go to admin area and define a nice wiki text here as a fixed skeletal for all projects.',
-    remove_help: 0,
-    remove_mypage: 0,
-    remove_news: 0
-  }
-
+  default_settings = Additionals.load_settings
   5.times do |i|
     default_settings['custom_menu' + i.to_s + '_name'] = ''
     default_settings['custom_menu' + i.to_s + '_url'] = ''
@@ -50,6 +23,8 @@ Redmine::Plugin.register :additionals do
 
   project_module :issue_tracking do
     permission :edit_closed_issues, {}
+    permission :edit_issue_author, {}
+    permission :change_new_issue_author, {}
   end
 
   project_module :time_tracking do