+++ lib/additionals.rb
@@ -1,28 +1,52 @@
-if ActiveRecord::Base.connection.table_exists?(:settings)
-  Rails.configuration.to_prepare do
-    module Additionals
-      MAX_CUSTOM_MENU_ITEMS = 5
-      def self.settings
-        ActionController::Parameters.new(Setting[:plugin_additionals])
+module Additionals
+  MAX_CUSTOM_MENU_ITEMS = 5
+
+  def self.settings
+    ActionController::Parameters.new(Setting[:plugin_additionals])
+  end
+
+  def self.incompatible_plugins(plugins = [], title = 'additionals')
+    plugins.each do |plugin|
+      if Redmine::Plugin.installed?(plugin)
+        raise "\n\033[31m#{title} plugin cannot be used with #{plugin} plugin'.\033[0m"
       end
     end
+  end
 
-    if Redmine::Plugin.installed?('redmine_tweaks')
-      raise "\n\033[31madditionals plugin cannot be used with redmine_tweaks plugin'.\033[0m"
+  def self.patch(patches = [], plugin_id = 'additionals')
+    patches.each do |name|
+      patch_dir = Rails.root.join('plugins', plugin_id, 'lib', plugin_id, 'patches')
+      require "#{patch_dir}/#{name.underscore}_patch"
+
+      target = name.constantize
+      patch = "#{plugin_id.camelize}::Patches::#{name}Patch".constantize
+
+      unless target.included_modules.include?(patch)
+        target.send(:include, patch)
+      end
     end
+  end
 
-    if Redmine::Plugin.installed?('common_libraries')
-      raise "\n\033[31madditionals plugin cannot be used with common_libraries plugin'.\033[0m"
+  def self.load_macros(macros = [], plugin_id = 'additionals')
+    macro_dir = Rails.root.join('plugins', plugin_id, 'lib', plugin_id, 'wiki_macros')
+    macros.each do |macro|
+      require_dependency "#{macro_dir}/#{macro.underscore}_macro"
     end
+  end
 
-    # Patches
-    # require_dependency 'additionals/patches/custom_help_url'
-    require_dependency 'additionals/patches/issue_patch'
-    require_dependency 'additionals/patches/time_entry_patch'
-    require_dependency 'additionals/patches/wiki_patch'
-    require_dependency 'additionals/patches/wiki_controller_patch'
-    require_dependency 'additionals/patches/wiki_pdf_helper_patch'
-    require 'additionals/patches/application_controller_patch'
+  def self.load_settings(plugin_id = 'additionals')
+    data = YAML.safe_load(ERB.new(IO.read(Rails.root.join('plugins',
+                                                          plugin_id,
+                                                          'config',
+                                                          'settings.yml'))).result) || {}
+    data.symbolize_keys
+  end
+end
+
+if ActiveRecord::Base.connection.table_exists?(:settings)
+  Rails.configuration.to_prepare do
+    Additionals.incompatible_plugins(%w[redmine_tweaks common_libraries redmine_editauthor redmine_changeauthor])
+    Additionals.patch(%w[Issue IssuesController TimeEntry Wiki WikiController ApplicationController])
 
     Rails.configuration.assets.paths << Emoji.images_path
     # Send Emoji Patches to all wiki formatters available to be able to switch formatter without app restart
@@ -35,30 +59,20 @@ if ActiveRecord::Base.connection.table_exists?(:settings)
       end
     end
 
+    # Static class patches
+    require_dependency 'additionals/patches/wiki_pdf_helper_patch'
+    require_dependency 'additionals/patches/access_control_patch'
+
     # Global helpers
     require_dependency 'additionals/helpers'
 
     # Hooks
     require_dependency 'additionals/hooks'
 
-    # Wiki macros
-    require_dependency 'additionals/wiki_macros/calendar'
-    require_dependency 'additionals/wiki_macros/cryptocompare'
-    require_dependency 'additionals/wiki_macros/date'
-    require_dependency 'additionals/wiki_macros/gist'
-    require_dependency 'additionals/wiki_macros/issue_macro'
-    require_dependency 'additionals/wiki_macros/last_updated_at'
-    require_dependency 'additionals/wiki_macros/last_updated_by'
-    require_dependency 'additionals/wiki_macros/member_macro'
-    require_dependency 'additionals/wiki_macros/project_macro'
-    require_dependency 'additionals/wiki_macros/recently_updated'
-    require_dependency 'additionals/wiki_macros/reddit'
-    require_dependency 'additionals/wiki_macros/slideshare'
-    require_dependency 'additionals/wiki_macros/tradingview'
-    require_dependency 'additionals/wiki_macros/twitter'
-    require_dependency 'additionals/wiki_macros/user_macro'
-    require_dependency 'additionals/wiki_macros/vimeo'
-    require_dependency 'additionals/wiki_macros/youtube'
+    # Macros
+    Additionals.load_macros(%w[calendar cryptocompare date gist issue last_updated_at last_updated_by
+                               member project recently_updated reddit slideshare tradingview
+                               twitter user vimeo youtube])
   end
 
   # include deface overwrites